+++ FoodForYou.Persistence.EntityFramework/Entities/Product.cs
@@ -10,6 +10,6 @@ namespace FoodForYou.Persistence.EntityFramework.Entities
         public int ProductId { get; set; }
         public string ProductName { get; set; }
 
-        public double UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
     }
 }
